JANITORIAL: add cast to avoid signedness mismatch in comparison

+++ engines/tot/engine.cpp
@@ -3967,7 +3967,7 @@ void TotEngine::sayLine(
 					charCounter -= 1;
 				} while (text.text[charCounter] != ' ');
 				lineBreaks[lineBreakCount] = charCounter + 1;
-			} while (charCounter + 1 <= ((int)text.text.size() - (int)width));
+			} while (((int)charCounter + 1) <= ((int)text.text.size() - (int)width));
 
 			lineBreakCount += 1;
 			lineBreaks[lineBreakCount] = text.text.size();
